tests/wpt/web-platform-tests/infrastructure/expected-fail/precondition-without-setup.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Precondition without wrapping setup</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
assert_precondition(false);
</script>
